you was conflicted

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -28,7 +28,7 @@ int count(int x)
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	int xor;
+	unsigned long int xor;
 
 	unsigned long int xor = n ^ m;
